API backend release

+++ app/Http/Controllers/BakerymathController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Bakerymath;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 
 class BakerymathController extends Controller
@@ -12,10 +13,10 @@ class BakerymathController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index($user)
+    public function index()
     {
         // should pass user api token parameter
-        $showBakerymath = Bakerymath::where('user_id',$user)->get();
+        $showBakerymath = Bakerymath::where('user_id', Auth::user()->id)->get();
         return $showBakerymath;
     }
 
@@ -38,11 +39,10 @@ class BakerymathController extends Controller
     public function store(Request $request)
     {
         $this->validate($request,[
-            'ingredient' => 'min:3|max:15|required|unique:bakerymaths',
+            'ingredient' => 'min:3|max:15|required|unique:bakerymaths,ingredient,NULL,id,user_id,'.Auth::user()->id,
             'quantity' => 'required|numeric',
             // not present on table * for checking only if flour is created in table
             'flour' => 'required',
-            // 'user_id' => 'required'
         ],[
             'flour.required' => 'Please add "flour" ingredient first!'
         ]);
@@ -51,7 +51,7 @@ class BakerymathController extends Controller
         $bakerymath->fill($request->all());
         $bakerymath->ingredient = $request->input('ingredient');
         $bakerymath->quantity = $request->input('quantity');
-        $bakerymath->user_id = 1; // $request->input('user_id');
+        $bakerymath->user_id = Auth::user()->id;
         // this should check for user api token
         // $bakerymath->user()->associate($user);
         $bakerymath->save();
@@ -91,7 +91,7 @@ class BakerymathController extends Controller
     public function update(Request $request, Bakerymath $bakerymath)
     {
         $this->validate($request,[
-            'ingredient' => 'required|unique:bakerymaths,ingredient,'.$bakerymath->id,
+            'ingredient' => 'required|unique:bakerymaths,ingredient,NULL,id,user_id,'.Auth::user()->id,
             'quantity' => 'numeric|min:0.1',
             // not present on table * for checking only if flour is created in table
             'flour' => 'required',
@@ -127,17 +127,9 @@ class BakerymathController extends Controller
         }
     }
 
-    public function sweep($user)
+    public function sweep()
     {
-        try {
-
-            $sweepBakerymath = Bakerymath::where('user_id',$user)->delete();
-            return response('Accepted', 202 )->header('Content-Type', 'application/json');
-
-        } catch (\Exception $e) {
-
-            return response('Connection Failure', 420)->header('Content-Type', 'application/json');
-
-        }
+        $sweepBakerymath = Bakerymath::where('user_id',Auth::user()->id)->delete();
+        return response('Accepted', 202 )->header('Content-Type', 'application/json');
     }
 }
